Update README.md
The link would take the user to the c code execution video and the second video would take him to the frontend html and css code for gym registration. since i dont know about the backend so the video is localised as of now but one can access the links.